SoftwareDeliveryMachine is now an interface
