+++ src/common/listener/RepoCreationListener.ts
@@ -28,6 +28,6 @@ export interface RepoCreationListenerInvocation extends RepoListenerInvocation {
 /**
  * Respond to the creation of a new repo.
  * Note that it may not have code in it, so you may want to use
- * a PushListener! See SoftwareDeliveryMachine.addNewRepoWithCodeActions
+ * a PushListener! See DefaultSoftwareDeliveryMachine.addNewRepoWithCodeActions
  */
 export type RepoCreationListener = SdmListener<RepoCreationListenerInvocation>;
